pyproject.toml
[project]
name = "criando-agentes-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agno>=2.2.6",
    "fastapi>=0.121.3",
    "groq>=0.33.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.13",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]
